moved EDI validation to its own class
